Fixed some headers, fixed broken switch statement in Robot

+++ src/Team167Lib/SwerveDrive.cpp
@@ -180,6 +180,6 @@ int SwerveDrive::GetWheelAngle()
 {
         float gyroangle = gyro->GetAngle();
         int robotRelativeWheelAngle = rotateTalon1->GetEncPosition() * 360 / rotateEncoderLines;
-        int fieldRelativeWheelAngle = (gyroangle + robotRelativeWheelAngle) % 360;
+        int fieldRelativeWheelAngle = ((int)gyroangle + robotRelativeWheelAngle) % 360;
         return fieldRelativeWheelAngle;
 }	//finds the angle the wheels are facing relative to the field
